Move configure_settings from utils to config.py
Add documentation for the various configuration options
