+++ docs/training.rst
@@ -306,94 +306,8 @@ some offline training method then those two processes would conflict
 with each other, likely resulting in unintended effects.  So NengoDL will
 assume that those elements should not be optimized.
 
-Any of these default behaviours can be overridden using `Nengo's config system
-<https://www.nengo.ai/nengo/config.html>`_.  Specifically, setting the
-``trainable`` config attribute for an object will control whether or not it
-will be optimized.
-
-:func:`.configure_settings` is a utility function that can be used to add a
-configurable ``trainable`` attribute to the objects in a network.  Setting
-``trainable=None`` will use the defaults described above, or True/False can
-be passed to override the default for all objects in a model.
-
-For example, suppose we only want to optimize one connection in our network,
-while leaving everything else unchanged.  This could be achieved via
-
-.. code-block:: python
-
-    with nengo.Network() as net:
-        # this adds the `trainable` attribute to all the trainable objects
-        # in the network, and initializes it to `False`
-        nengo_dl.configure_settings(trainable=False)
-
-        a = nengo.Node([0])
-        b = nengo.Ensemble(10, 1)
-        c = nengo.Node(size_in=1)
-
-        nengo.Connection(a, b)
-
-        # make this specific connection trainable
-        conn = nengo.Connection(b, c)
-        net.config[conn].trainable = True
-
-Or if we wanted to disable training for some subnetwork:
-
-.. code-block:: python
-
-    with nengo.Network() as net:
-        nengo_dl.configure_settings(trainable=None)
-        ...
-        with nengo.Network() as subnet:
-            net.config[subnet].trainable = False
-            ...
-
-Note that ``config[nengo.Ensemble].trainable`` controls both encoders and
-biases, as both are properties of an Ensemble.  However, it is possible to
-separately control the biases via ``config[nengo.ensemble.Neurons].trainable``
-or ``config[my_ensemble.neurons].trainable``.
-
-There are two important caveats to keep in mind when configuring ``trainable``,
-which differ from the standard config behaviour:
-
-1. ``trainable`` applies to all objects in a network, regardless of whether
-   they were created before or after ``trainable`` is set.  For example,
-
-   .. code-block:: python
-
-       with nengo.Network() as net:
-           ...
-           net.config[nengo.Ensemble].trainable = False
-           a = nengo.Ensemble(10, 1)
-           ...
-
-   is the same as
-
-   .. code-block:: python
-
-       with nengo.Network() as net:
-           ...
-           a = nengo.Ensemble(10, 1)
-           net.config[nengo.Ensemble].trainable = False
-           ...
-
-
-2. ``trainable`` can only be set on the config of the top-level network.  For
-   example,
-
-   .. code-block:: python
-
-       with nengo.Network() as net:
-           nengo_dl.configure_settings(trainable=None)
-
-           with nengo.Network() as subnet:
-               my_ens = nengo.Ensemble(...)
-
-               # incorrect
-               subnet.config[my_ens].trainable = False
-
-               # correct
-               net.config[my_ens].trainable = False
-
+Any of these default behaviours can be overridden using the
+:ref:`"trainable" config option <config-trainable>`.
 
 Saving and loading parameters
 -----------------------------
