+++ nengo_dl/__init__.py
@@ -10,7 +10,7 @@ from nengo_dl import (
 from nengo_dl import dists
 from nengo_dl.simulator import Simulator
 from nengo_dl.tensor_node import TensorNode, tensor_layer, reshaped
-from nengo_dl.utils import configure_settings
+from nengo_dl.config import configure_settings
 from nengo_dl.neurons import SoftLIFRate
 
 # apply tensorflow monkey patches
